NewsData/API/requirements.txt
flask==3.1.1
flask_cors==6.0.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0
gunicorn==21.2.0
python-dotenv==1.0.0